+++ cdap-standalone/src/main/resources/logback.xml
@@ -39,6 +39,7 @@
   <logger name="org.quartz.core" level="WARN"/>
   <logger name="org.eclipse.jetty" level="WARN"/>
   <logger name="io.netty.util.internal" level="WARN"/>
+  <logger name="co.cask.cdap.logging.appender.system.LogFileManager" level="ERROR"/>
   <logger name="co.cask.cdap.operations.OperationalStats" level="ERROR"/>
   <logger name="co.cask.cdap.extension.AbstractExtensionLoader" level="ERROR"/>
