+++ cdap-watchdog/src/main/java/co/cask/cdap/logging/appender/system/LogFileOutputStream.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package co.cask.cdap.logging.framework;
+package co.cask.cdap.logging.appender.system;
 
 import ch.qos.logback.classic.spi.ILoggingEvent;
 import co.cask.cdap.common.io.ByteBuffers;
