+++ cdap-watchdog/src/main/resources/cdap-log-pipeline.xml
@@ -24,7 +24,7 @@
   <!-- This is for backward compatibility for <= 4.0, since it was hardcoded to Bytes.toBytes(100). Don't modify. -->
   <property name="log.process.pipeline.checkpoint.prefix.num" value="100"/>
 
-  <appender name="CDAP" class="co.cask.cdap.logging.framework.CDAPLogAppender">
+  <appender name="CDAP" class="co.cask.cdap.logging.appender.system.CDAPLogAppender">
     <dirPermissions>${dir.permissions}</dirPermissions>
     <filePermissions>${file.permissions}</filePermissions>
     <syncIntervalBytes>${file.sync.interval.bytes}</syncIntervalBytes>
@@ -35,7 +35,7 @@
     <fileCleanupTransactionTimeout>${file.cleanup.transaction.timeout}</fileCleanupTransactionTimeout>
   </appender>
 
-  <appender name="METRICS" class="co.cask.cdap.logging.framework.MetricsLogAppender"/>
+  <appender name="METRICS" class="co.cask.cdap.logging.appender.system.MetricsLogAppender"/>
 
   <root level="ALL">
     <appender-ref ref="CDAP"/>
