+++ cdap-watchdog/src/test/java/co/cask/cdap/logging/framework/distributed/DistributedLogFrameworkTest.java
@@ -41,8 +41,8 @@ import co.cask.cdap.data.runtime.SystemDatasetRuntimeModule;
 import co.cask.cdap.data.runtime.TransactionExecutorModule;
 import co.cask.cdap.kafka.KafkaTester;
 import co.cask.cdap.logging.appender.LogMessage;
+import co.cask.cdap.logging.appender.system.LogPathIdentifier;
 import co.cask.cdap.logging.filter.Filter;
-import co.cask.cdap.logging.framework.LogPathIdentifier;
 import co.cask.cdap.logging.guice.DistributedLogFrameworkModule;
 import co.cask.cdap.logging.meta.Checkpoint;
 import co.cask.cdap.logging.meta.CheckpointManagerFactory;
