+++ cdap-watchdog/src/test/java/co/cask/cdap/logging/read/FileMetadataTest.java
@@ -37,9 +37,9 @@ import co.cask.cdap.data2.dataset2.DatasetFramework;
 import co.cask.cdap.data2.dataset2.MultiThreadDatasetCache;
 import co.cask.cdap.data2.transaction.Transactions;
 import co.cask.cdap.logging.LoggingConfiguration;
+import co.cask.cdap.logging.appender.system.CDAPLogAppender;
+import co.cask.cdap.logging.appender.system.LogPathIdentifier;
 import co.cask.cdap.logging.context.LoggingContextHelper;
-import co.cask.cdap.logging.framework.CDAPLogAppender;
-import co.cask.cdap.logging.framework.LogPathIdentifier;
 import co.cask.cdap.logging.guice.LoggingModules;
 import co.cask.cdap.logging.meta.FileMetaDataReader;
 import co.cask.cdap.logging.meta.FileMetaDataWriter;
